requirements.txt
# AI News Tracker - 依赖包列表

# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0
pydantic>=2.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0

# 数据采集
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
aiohttp>=3.9.0

# AI/LLM
openai>=1.3.0
tiktoken>=0.5.0

# 任务调度
apscheduler>=3.10.0

# 工具库
python-dotenv>=1.0.0
click>=8.1.0

# 飞书SDK (可选)
lark-oapi>=1.2.0

# 其他
feedgen>=0.9.0
markdownify>=0.11.0
